pyproject.toml
[tool.poetry]
name = "explorelimits"
version = "0.1.0"
description = ""
authors = ["Mahendra <mahendra@taiyo.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
